Extract run_summary.py from pipeline.py; emit lexical retrieval block
