+++ examples/querygym_pyserini/pipeline.py
@@ -38,102 +38,8 @@ from examples.querygym_pyserini.utils import (
 from examples.querygym_pyserini import reformulate_queries
 from examples.querygym_pyserini import retrieve
 from examples.querygym_pyserini import evaluate
-from reproducibility.lib import build_run_summary, validate, ValidationError
-
-
-def _load_dataset_config_from_registry(dataset_id: str, registry_path: str) -> dict | None:
-    """Pull dataset_config fields from dataset_registry.yaml. None if not registered."""
-    try:
-        import yaml
-        with open(registry_path, 'r') as f:
-            registry = yaml.safe_load(f) or {}
-    except Exception:
-        return None
-    entry = registry.get('datasets', {}).get(dataset_id)
-    if not entry:
-        return None
-    return {
-        'topics': entry.get('topics', {}).get('name', ''),
-        'index': entry.get('index', {}).get('name', ''),
-        'num_queries': 0,  # filled by reformulation metadata when available
-        'bm25_weights': entry.get('bm25_weights', {'k1': 0.0, 'b': 0.0}),
-    }
-
-
-def _build_v1_summary(
-    *, results, dataset_name, method, model, method_params, llm_config,
-    steps, pipeline_time, registry_path, queries_file, index_name,
-) -> dict:
-    """Pull fields from per-step metadata and call reproducibility.lib.build_run_summary."""
-    reform = results.get('reformulation', {})
-    retrieval = results.get('retrieval', {})
-    evaluation = results.get('evaluation', {})
-
-    reform_inner = reform.get('reformulation', {}) if isinstance(reform, dict) else {}
-    dataset_inner = reform.get('dataset', {}) if isinstance(reform, dict) else {}
-
-    # Resolve dataset_config: prefer reformulation metadata (richest), fall back
-    # to the registry, then to file-based info.
-    if dataset_inner.get('topics') or dataset_inner.get('index'):
-        dataset_config = {
-            'topics': dataset_inner.get('topics') or '',
-            'index': dataset_inner.get('index') or (index_name or ''),
-            'num_queries': int(dataset_inner.get('num_queries') or 0),
-            'bm25_weights': dataset_inner.get('bm25_weights') or {'k1': 0.0, 'b': 0.0},
-        }
-    else:
-        dataset_config = _load_dataset_config_from_registry(dataset_name, registry_path) or {
-            'topics': '',
-            'index': index_name or '',
-            'num_queries': 0,
-            'bm25_weights': {'k1': 0.0, 'b': 0.0},
-        }
-        # If reformulation produced num_queries but no other config, splice it in.
-        if dataset_inner.get('num_queries'):
-            dataset_config['num_queries'] = int(dataset_inner['num_queries'])
-
-    # Searcher: use reformulation's searcher info if present, otherwise unknown.
-    searcher_info = reform_inner.get('searcher') or {}
-    searcher = {
-        'name': searcher_info.get('name') or 'unknown',
-        'type': searcher_info.get('type') or 'unknown',
-    }
-
-    # Effective method_params / llm_config: prefer the per-step metadata's view
-    # (already resolved with method-specific defaults), fall back to caller args.
-    eff_method_params = reform_inner.get('method_params') or dict(method_params or {})
-    # method_params may include a non-serializable searcher; strip it.
-    eff_method_params = {
-        k: v for k, v in eff_method_params.items()
-        if not callable(v) and not hasattr(v, '__dict__') or isinstance(v, (dict, list, str, int, float, bool, type(None)))
-    }
-    eff_llm_config = reform_inner.get('llm_config') or dict(llm_config or {})
-    # Keep only schema-relevant keys; build_run_summary accepts extras but the schema
-    # only requires temperature + max_tokens.
-    if 'temperature' not in eff_llm_config:
-        eff_llm_config['temperature'] = (llm_config or {}).get('temperature', 0.0)
-    if 'max_tokens' not in eff_llm_config:
-        eff_llm_config['max_tokens'] = (llm_config or {}).get('max_tokens', 1)
-
-    timing = {
-        'reformulation_seconds': float(reform.get('timing', {}).get('total_time_seconds', 0.0)),
-        'retrieval_seconds': float(retrieval.get('timing', {}).get('total_time_seconds', 0.0)),
-        'evaluation_seconds': float(evaluation.get('timing', {}).get('eval_time_seconds', 0.0)),
-    }
-
-    return build_run_summary(
-        dataset_id=dataset_name or queries_file.stem if queries_file else (dataset_name or 'unknown'),
-        method_id=method,
-        model=model,
-        method_params=eff_method_params,
-        llm_config=eff_llm_config,
-        searcher=searcher,
-        dataset_config=dataset_config,
-        metrics=evaluation.get('results', {}),
-        timing=timing,
-        steps_completed=steps,
-        total_time_seconds=pipeline_time,
-    )
+from reproducibility.lib import validate, ValidationError
+from examples.querygym_pyserini.run_summary import _build_v1_summary
 
 
 def run_pipeline(
